\name{gDefrag-package}
\alias{gDefrag-package}
\alias{gDefrag}
\docType{package}
\title{
\packageTitle{gDefrag}
}
\description{
\packageDescription{gDefrag}
}
\details{
Our approach is based on graph-based tools, and relies only on the spatial configuration of the infrastructures and the layer of the resource to be evaluated (e.g. habitat area or quality). The package interprets the landscape as a graph-like structure, taking advantage of the structural simplicity of this approach. The main assumption is that connecting a larger amount of a given resource is benefit for the species persistence, following the concept of 'accessible habitat' by Eigenbrod et al. (2008). It then prioritizes the sections that have higher impact in overall connectivity according to three methods, including the Integral Index of Connectivity (IIC) developed by Pascual-Hortal and Saura (2006). It can be used to assist in the de-fragmentation process caused by any linear infrastructure type, including roads, railways, power lines, pipelines, water channels, and fences. Hereafter we will refer to these simply as 'roads'.      
}
\author{
\packageAuthor{gDefrag}

Maintainer: \packageMaintainer{gDefrag}
}
\references{
Eigenbrod, F., Hecnar, S.J., Fahrig, L. (2008). Accessible habitat: an improved measure of the effects of habitat loss and roads on wildlife populations. Landscape Ecology, 23(2), 159-168.

Grilo, C., Ascensao, F., Santos-Reis, M., Bissonette, J.A. (2011). Do well-connected landscapes promote road-related mortality?. European Journal of Wildlife Research, 57(4), 707-716.

Pascual-Hortal, L., Saura, S. (2006). Comparison and development of new graph-based landscape connectivity indices: towards the priorization of habitat patches and corridors for conservation. Landscape Ecology, 21(7), 959-967.

Santos, S.M., Marques, J.T., Lourenco, A., Medinas, D., Barbosa, A.M., Beja, P., Mira, A. (2015). Sampling effects on the identification of roadkill hotspots: implications for survey design. Journal of Environmental Management, 162, 87-95.
}
